+++ test/namedtuple.jl
@@ -276,6 +276,7 @@ abstr_nt_22194_3()
 @test Base.delete((a=1, b=2), :a) == (b=2,)
 @test Base.delete((a=1, b=2, c=3), :b) == (a=1, c=3)
 @test Base.delete((a=1, b=2, c=3), :z) == (a=1, b=2, c=3)
+@test Base.delete((a=1, b=2, c=3), (:b, :a)) == (c=3,)
 
 @test Base.structdiff((a=1, b=2), (b=3,)) == (a=1,)
 @test Base.structdiff((a=1, b=2, z=20), (b=3,)) == (a=1, z=20)